@@ -1,5 +1,5 @@
 {
-  "name": "alpha-foundation-build",
+  "name": "sqlchef",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -7,6 +7,7 @@
     "build:dev": "./node_modules/.bin/webpack",
     "build:prod": "./node_modules/.bin/webpack",
     "dev": "npm run build:dev && node server.js",
+    "prod": "npm run build:prod && node server.js",
     "stats": "./node_modules/.bin/webpack --profile --json > stats.json",
     "test": "./node_modules/.bin/jest"
   },
@@ -16,26 +17,34 @@
     "typescript",
     "react",
     "scss",
-    "hmr"
+    "hmr",
+    "styled components",
+    "phaser",
+    "phaser 3"
   ],
   "author": "Jeff Peoples",
   "license": "MIT",
   "dependencies": {
     "express": "^4.16.4",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "phaser": "^3.15.1",
+    "react": "16.7.0-alpha.0 - next",
+    "react-dom": "16.7.0-alpha.0 - next",
+    "react-vis": "^1.11.5",
+    "styled-bootstrap-components": "^3.0.1",
     "styled-components": "^4.1.3"
   },
   "devDependencies": {
-    "@types/styled-components": "^4.1.4",
     "@types/jest": "^23.3.10",
     "@types/react": "^16.7.17",
     "@types/react-dom": "^16.0.11",
+    "@types/styled-components": "^4.1.4",
     "css-loader": "^2.0.1",
+    "expose-loader": "^0.7.5",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^23.6.0",
     "jest-dom": "^3.0.0",
     "node-sass": "^4.11.0",
+    "react-test-renderer": "^16.7.0",
     "react-testing-library": "^5.4.0",
     "sass-loader": "^7.1.0",
     "source-map-loader": "^0.2.4",
