@@ -4,18 +4,31 @@ const webpack = require('webpack')
 const webpackConfig = require('./webpack.config.js')
 const app = express()
 const port = process.env.PORT || 3000;
+const environment = process.env.npm_lifecycle_event; 
 
+
+// Routes -- START
 app.get('/', (req, res) => {
    res.sendFile(path.resolve(__dirname, 'dist', 'index.html'));
 });
 
-let compiler = webpack(webpackConfig);
+// Routes -- END
 
-app.use(require('webpack-dev-middleware')(compiler, {
-   noInfo: true, publicPath: webpackConfig.output.publicPath, stats:    { colors: true }
-}));
-app.use(require('webpack-hot-middleware')(compiler));
+
+// Webpack Development Middleware -- START
+if(environment === "dev"){
+
+   let compiler = webpack(webpackConfig);
+
+   app.use(require('webpack-dev-middleware')(compiler, {
+      noInfo: true, publicPath: webpackConfig.output.publicPath, stats:    { colors: true }
+   }));
+
+   app.use(require('webpack-hot-middleware')(compiler));
+}
+
+// Webpack Development Middleware -- END
+
 app.use(express.static(path.resolve(__dirname,'dist')));
-
 
 app.listen(port, () => { console.log(`App is listening on port ${port}`) });
